macos: Enable secure state restoration.
Beginning with macOS 12 Monterey, a new method of serializing window
state was introduced. The old method is vulnerable to process injection
attacks. However, due to compatibility concerns, the old method is still
the default and applications must opt in to the new method. They do so
by responding positively to the app delegate selector
`applicationSupportsSecureRestorableState:`.

https://developer.apple.com/documentation/appkit/nsapplicationdelegate/3762521-applicationsupportssecurerestora